refactor ♻️: Rename prompts.py to commit_prompts.py in AI subpackage
This change renames the `prompts.py` file to `commit_prompts.py` within the AI subpackage. The renaming is part of a broader refactoring effort that aims to centralize system prompts management, improving code organization and maintainability. The change affects both the documentation and the actual implementation files, ensuring consistency across the project.
