+++ docs/release-notes.md
@@ -19,7 +19,7 @@ This release introduces significant architectural improvements and refactoring a
 - Reorganized AI functionality into focused components:
   - `client.py`: Improved AI client with better request handling
   - `generation.py`: Streamlined commit message generation
-  - `prompts.py`: Centralized system prompts management
+  - `commit_prompts.py`: Centralized system prompts management
 
 ### Command Line Interface 🖥️
